requirements.txt
Package             Version
------------------- -----------
aiohttp             3.9.3
aiosignal           1.3.1
annotated-types     0.6.0
anyio               4.3.0
argcomplete         2.0.0
asttokens           2.4.1
attrs               23.2.0
azure-core          1.30.0
azure-identity      1.15.0
Beaker              1.12.1
beautifulsoup4      4.12.3
blivet              3.8.2
blivet-gui          2.5.0
Brlapi              0.8.5
certifi             2024.2.2
cffi                1.15.1
charset-normalizer  3.2.0
click               8.1.3
comm                0.2.1
contourpy           1.2.0
cryptography        41.0.7
cupshelpers         1.0
cycler              0.12.1
dasbus              1.7
dataclasses-json    0.6.4
dbus-python         1.3.2
debugpy             1.8.1
decorator           5.1.1
distro              1.8.0
dnf                 4.19.0
docopt              0.6.2
executing           2.0.1
fedora-third-party  0.10
file-magic          0.4.0
fonttools           4.49.0
fros                1.1
frozenlist          1.4.1
greenlet            3.0.3
h11                 0.14.0
httpcore            1.0.3
httpx               0.26.0
humanize            3.13.1
idna                3.4
install             1.3.5
ipykernel           6.29.2
ipython             8.21.0
jedi                0.19.1
joblib              1.3.2
jsonpatch           1.33
jsonpointer         2.4
jupyter_client      8.6.0
jupyter_core        5.7.1
kiwisolver          1.4.5
langchain           0.1.7
langchain-community 0.0.20
langchain-core      0.1.23
langchain-openai    0.0.6
langsmith           0.0.87
langtable           0.0.64
libcomps            0.1.20
libdnf              0.73.0
lxml                4.9.3
Mako                1.2.3
MarkupSafe          2.1.3
marshmallow         3.20.2
matplotlib          3.8.3
matplotlib-inline   0.1.6
msal                1.26.0
msal-extensions     1.1.0
multidict           6.0.5
mypy-extensions     1.0.0
nest-asyncio        1.6.0
nftables            0.1
num2words           0.5.13
numpy               1.26.4
olefile             0.46
openai              1.12.0
packaging           23.2
pandas              2.2.0
parso               0.8.3
Paste               3.5.3
pexpect             4.8.0
pid                 2.2.3
pillow              10.2.0
pip                 23.2.1
platformdirs        4.2.0
plotly              5.19.0
ply                 3.11
portalocker         2.8.2
productmd           1.38
prompt-toolkit      3.0.43
psutil              5.9.8
ptyprocess          0.7.0
pure-eval           0.2.2
pwquality           1.4.5
pycairo             1.25.1
pycparser           2.20
pycrypto            2.6.1
pycups              2.0.1
pycurl              7.45.2
pydantic            2.6.1
pydantic_core       2.16.2
pyenchant           3.2.2
Pygments            2.17.2
PyGObject           3.46.0
PyJWT               2.8.0
pykickstart         3.48
pyOpenSSL           23.2.0
pyparsing           3.1.1
pyparted            3.13.0
PySocks             1.7.1
python-augeas       1.1.0
python-dateutil     2.8.2
python-dotenv       1.0.1
python-meh          0.51
pytz                2024.1
pyudev              0.24.1
pyxdg               0.27
PyYAML              6.0.1
pyzmq               25.1.2
regex               2023.12.25
requests            2.28.2
requests-file       1.5.1
requests-ftp        0.3.1
rpm                 4.19.1.1
scikit-learn        1.4.1.post1
scipy               1.12.0
selinux             3.5
sepolicy            3.5
setools             4.4.3
setuptools          67.7.2
simpleaudio         1.0.4
simpleline          1.9.0
six                 1.16.0
sniffio             1.3.0
sos                 4.6.0
soupsieve           2.5
SQLAlchemy          2.0.27
stack-data          0.6.3
systemd-python      235
Tempita             0.5.2
tenacity            8.2.3
threadpoolctl       3.3.0
tiktoken            0.6.0
tornado             6.4
tqdm                4.66.2
traitlets           5.14.1
typing_extensions   4.9.0
typing-inspect      0.9.0
tzdata              2024.1
urllib3             1.26.18
wcwidth             0.2.13
yarl                1.9.4
